Remove photos repo and add a post repo
